third_party/WebKit/LayoutTests/fast/css/quirk-inherit-ua-only.html

<!DOCTYPE html>
<title>The '-internal-quirk-inherit' CSS value is only exposed to UA stylesheets</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
#t1 {
  color: blue;
  color: -internal-quirk-inherit;
}
</style>
<table id="t1"></table>
<table style="color: -internal-quirk-inherit"></table>
<script>
test(function() {
  var tables = document.querySelectorAll('table');
  assert_equals(getComputedStyle(tables[0]).color, 'rgb(0, 0, 255)');
  assert_equals(tables[1].style.color, '');
});
</script>
